fix the clone/edit/delete in the main tab
